Extend the GifMaker

+++ lipstick/animations.py
@@ -3,10 +3,10 @@ import glob
 import shutil
 import time
 from pathlib import Path
-from PIL import Image
 from IPython import display
-import numpy as np
 import matplotlib.pyplot as plt
+import numpy as np
+import imageio
 
 
 def update_fig(fig, axes, sleep=0):
@@ -25,45 +25,52 @@ def update_fig(fig, axes, sleep=0):
 
 
 class GifMaker:
-    def __init__(self, filename, fps=30.0):
+    def __init__(self, filename, fps=30.0, dpi='figure', temp_dir_name="gifmaker_cache", keep_dir=False,):
         """
         Saves gif from added figures.
 
         Args:
             filename (str): full path and filename of the resulting gif (e.g. "sample.gif")
             fps (float, optional): Frames per second. Defaults to 30.0.
+            dpi (int, optional): Dots per inch of each image. Defaults to the dpi setting of the matplotlib figure which is being passed. Choose 150 for high quality gifs.
+            temp_dir_name (str, optional): Name of the temporary directory. Defaults to "gifmaker_cache".
+            keep_dir (bool, optional): If True, the temporary directory is not deleted. Useful when trying to create videos with custom ffmpeg commands. Defaults to False.
+
+        How to convert the gif to a video:
+            - set keep_dir=True
+            - after all the images are saved, navigate to the temporary directory
+            - make sure ffmpeg is installed on your system
+            - run the following command within the temporary directory:
+
+                ffmpeg -framerate 30 -pattern_type glob -i '*.png' -c:v libx264 out.mp4
+
+            - the video will be saved as "out.mp4" in the temporary directory, with 30 fps.
         """
 
         self.path = "/".join(filename.split(".")[0].split("/")[:-1])
         self.path = self.path if self.path else os.getcwd()
         self.name = filename.split(".")[0].split("/")[-1]
         self.fps = fps
-        self.temp_dir = "gifmaker_cache"
+        self.dpi = dpi
+        self.temp_dir = temp_dir_name
         self.i = 0
+        self.keep_dir = keep_dir
 
     def add(self, fig):
         fig.savefig(
             f"{self.temp_dir}/{self.name}" + str(self.i).zfill(4) + ".png",
             bbox_inches="tight",
+            dpi=self.dpi,
         )
         plt.close()
         self.i += 1
 
     def save(self):
-
         fp_in = "{}/{}*.png".format(self.temp_dir, self.name)
         fp_out = "{}/{}.gif".format(self.path, self.name)
 
-        imgs = [Image.open(f) for f in sorted(glob.glob(fp_in))]
-
-        imgs[0].save(
-            fp=fp_out,
-            format="GIF",
-            append_images=imgs[1:],
-            save_all=True,
-            duration=1000 / self.fps,
-            loop=0,
-        )
+        imgs = [imageio.imread(f) for f in sorted(glob.glob(fp_in))]
+        imageio.mimsave(fp_out, imgs, 'GIF', fps=self.fps)
 
     def show(self):
         fp_out = "{}/{}.gif".format(self.path, self.name)
@@ -77,4 +84,5 @@ class GifMaker:
 
     def __exit__(self, *kw):
         self.save()
-        shutil.rmtree(self.temp_dir)
+        if not self.keep_dir:
+            shutil.rmtree(self.temp_dir)
